[Tab Group UI] Lazily initialize
Lazily create the TabGroupUi for the bottom tab strip only when the
current activity tab is part of a tab group.

This avoids doing work when the selected tab at startup is not part of
a tab group.

Low-Coverage-Reason: COVERAGE_UNDERREPORTED TabManagementDelegateProvider has coverage in TabGroupUiOneshotSupplierUnitTest
Fixed: 351026325
Change-Id: I5d179f37aae9cfed3e665ce5fdb985bc7ffa0a18
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5675976
Reviewed-by: Sky Malice <skym@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1324891}