Move initialization of extension managers
Moves some of the lazily created extension managers to happen during
session creation, where the features can be disabled and/or the session
can be rejected as appropriate. Note that there are still other features
(Notably hand tracking and reference spaces) that this does not apply
to yet. Those will be addressed in a more general future change.

Change-Id: I5e53341756fccf3a76c70ab64b35dc4c99121d06
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5673513
Reviewed-by: Brandon Jones <bajones@chromium.org>
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1324995}