[Permissions] Origins in prompt title should be bold
This CL added the possibility to highlight a part of Label with
bold font. This is needed because we need to make an origin in a
permission prompt with bold font and we want to use a character
brake functionality which is available only in Label.

Bug: 40095827
Change-Id: I171c25add9ab09df0344e34a7b76255797b73ccb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5650766
Reviewed-by: Allen Bauer <kylixrd@chromium.org>
Commit-Queue: Yifan Luo <lyf@chromium.org>
Reviewed-by: Elias Klim <elklm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1325025}