Move ProductSpecificationsSet from bridge to service for delete
Moves ProductSpecificationsSet from the bridge to the service so that
the bridge focuses primarily on store/sync operations and the service
focuses on converting a ProductSpecificationsSet to specifics, thereby
working towards the goal of the bridge not needing to know about a
ProductSpecificationsSet. This separation between the service and the
bridge makes way for a simpler multi specifics implementation. This
is part of a number of changes and focuses on the delete operation.

Bug: 350345285
Change-Id: Id61288f5c090638208d9104f5650c659f7ff9358
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5675478
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Commit-Queue: David Maunder <davidjm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1325028}