[MSan] Fix use-after-dtor error in TabStripScrollContainer
TabStripScrollContainer registers itself as a ViewObserver of the
TabStrip but never unregisters itself. TabStrip is also an owned
descendent view of TabStripScrollContainer. This results in a
use-after-dtor when the TabStripScrollContainer is destroyed, since:

- base classes are destroyed in reverse order
- so ViewObserver is destroyed before View
- destroying View destroys its descendents
- destroying the descendent TabSTrip notifies the ViewObserver, but the
  ViewObserver is already destroyed

The fix is to use ScopedObservation to manage the observer lifetime; the
observer will be unregistered before the View base class runs.

In theory, an alternate fix would be to reorder the base classes, but it
would be pointless. By definition, TabStripScrollContainer's destructor
has already completed by the time the base class View destructor runs—so
the observations would end up at the no-op ViewObserver base class
implementation rather than being routed to TabStripScrollContainer.

Bug: 40222690
Change-Id: I776a224304eb0a38b180385eb64953c5342890c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5679614
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Allen Bauer <kylixrd@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1325051}