[Apps] Fix dependency issue in app termination observation
The AppTerminationObserver is a KeyedService that monitors for
app termination. However, when it is notified that the app is
terminating, it passes along a BrowserContext, which is then used
to retrieve other BrowserContext services. This may be unsafe if
the shutdown process for the BrowserContext has started, since
the context may be marked as dead, which results in a failed
assert.

Fix this by nulling out the termination observation when the
shutdown for KeyedServices begins for the associated BrowserContext.
There was already no guarantee that app termination would be called
for a BrowserContext if it had gone away, so this should, in theory,
be fine.

Bug: 352003806
Change-Id: Iba87a99aa45eebe62e0d30cd6a6d907f44dd02d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5688864
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Commit-Queue: Devlin Cronin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1325072}