Clarify some transforms used by OpenXR Anchors
This started as an investigation as to why the ANCHORS feature needed
the stage space to exist. It turns out that it only needs the stage
space if we're asking for an anchor to be created from the local-floor
reference space, which means that the existing entry for local-floor
is sufficient as it is essentially a transtive dependency. If we ever
switched exclusively to something like e.g. XR_EXT_LOCAL_FLOOR, this
dependency would not be present anymore either.

This updates some confusing comments in OpenXrAnchorManager and
OpenXrRenderLoop that are used in generating some of the transforms to
be more explicit in local terms about the offsets that are being
applied. Matrix multiplications were also added to explicitly account
for the current assumptions of which spaces are identity to ease future
refactors that would intend to remove those assumptions.

As the only code changes are additions of multiplication by identity,
and in order to request anchors in a way that needs local-floor, the
local-floor feature needs to also be enabled, this change should be a
no-op.

Change-Id: If4565efc2f535fc3b85659d041bdebd7d53f8063
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5672906
Reviewed-by: Piotr Bialecki <bialpio@chromium.org>
Commit-Queue: Piotr Bialecki <bialpio@chromium.org>
Auto-Submit: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1325145}