Add return callback to SwitchBiometricAuthBeforeFillingState
This CL modifies an existing passwords private delegate method
responsible for authentication and enabling biometric auth before
filling, by adding a return callback that is used to pass the result
of the authentication to the webui. It will be used in a followup CL
to show the confirmation toast.

Bug: 1512579
Change-Id: Ib04940c00a93a1945116ae57e0960afa282afeac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5625773
Reviewed-by: Tim <tjudkins@chromium.org>
Commit-Queue: Karol Sygiet <sygiet@google.com>
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1325344}