Rename NotifyDelegateOfInputEventPreDispatch to disambiguate meaning of delegate.
NotifyDelegateOfInputEventPreDispatch exists in
RenderInputRouterDelegate interface and any class implementing the
RIRDelegate interface would implement this method. And reading this
method in that class's context might be ambiguous, which is the case for
RenderWidgetHostImpl which has it's own delegate(WebContents). Renaming
this method to remove delegate mention.

Bug: b/347681162
Change-Id: I885c34050df3379f992f6a2cdd1e6901c2a05437
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5688592
Reviewed-by: Alexander Timin <altimin@chromium.org>
Reviewed-by: Stephen Nusko <nuskos@chromium.org>
Commit-Queue: Kartar Singh <kartarsingh@google.com>
Reviewed-by: Robert Flack <flackr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1325378}