[UPMClearMigratedData] Stop using AndroidBackendWithDoubleDeletion
If the login data is deleted, there is no need to create
AndroidBackendWithDoubleDeletion because the Android backend can be
used directly.

AndroidBackendWithDoubleDeletion is only created on startup when the
state pref is On so the migration has happened in the past. That will be
the same startup in which we attempt to delete the login data files.
It would only make sense to still have this backend if the login data
files deletion failed, but that should happen very rarely and it should
be corrected on the next startup.

Unless the deletion metrics show otherwise, it will be safe to delete
AndroidBackendWithDoubleDeletion when all the users have migrated and
deleted the login database files.

Bug: 350516053
Change-Id: I34c124f65d29c6f197f79d8d922327b93086317b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5688850
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Ivana Žužić <izuzic@google.com>
Cr-Commit-Position: refs/heads/main@{#1325385}