Add authentication logic to screenlock promo in pwm settings
This CL introduces an authentication logic when user interacts with
the screenlock promo card in settings. After the successful
authentication the pref responsible for biometric reauth is turned on.

Additionally this CL introduces a new confirmation toast that is
displayed after the successful authentication.

Bug: 1512579
Change-Id: I6e4f9926cb06926f26f2a376891a1b5b736e5a8e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5670372
Commit-Queue: Karol Sygiet <sygiet@google.com>
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1325387}