Add mojo type for javascript time based on unix epoch
Commonly, a time stamp needs to be sent from JS (e.g. to record a metric
for time-to-load). This can be confusing for developers because
javascript time is based on the Unix epoch, while C++ uses the Windows
epoch. This new type is a convenience method meant to clarify the
intentions of the value.

Bug: 350457796
Change-Id: I2c5a420e77718ec0e813ccfdd48c27a794757893
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5670453
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Joe Mason <joenotcharles@google.com>
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Commit-Queue: Mickey Burks <mickeyburks@chromium.org>
Reviewed-by: Oksana Zhuravlova <oksamyt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1325393}