Update Service Worker tests to work cross-process, cross-origin.
This CL modifies two ServiceWorkerDisableWebSecurityTests,
UnregisterNoCrash and UpdateNoCrash, to work with cross-process
cross-origin frames. Previously, the JS in the test files assumed
that the parent could directly manipulate objects in the child's
contentWindow, but this fails if the frames are cross-process to
one another. This CL modifies the tests to use postMessage for all
communications between the child and parent.

This modified version of the tests works regardless of the
process assignments of the two frames.

Bug: 40259221
Change-Id: Idfc3ab76275ff73d8a2da07de15f72dd6f177ec0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5689772
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Commit-Queue: James Maclean <wjmaclean@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1325406}