[ios] Unconditionally register KeyedService factories
The factories needs to be registered before the first BrowserState
is created (in order to allow the KeyedServiceTemplateFactory to
check that the BrowserState has been properly initialized before
a KeyedService is created).

Not registering the factory would not have prevented calling the
factory GetForBrowserState(...) method nor the creation of the
service. So this was incorrect.

Change EnsureBrowserStateKeyedServiceFactoriesBuilt to register
the factories unconditionally, and change the factories to return
nullptr if the feature is disabled.

Bug: 40158018
Change-Id: I4850a71e18c298ed1277f0dbf95241fcc6ae6be2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5683464
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1325409}