[ios] Convert targets from `static_library` to `source_set`
There is no good reason for those targets to be `static_library`
but this causes the build to be serialized until the library is
created building the object files in parallel and linking them
all in the final binary.

Convert the targets to `source_set`.

Bug: none
Change-Id: I13630d7039634b2db15bf024dd333d154be9b9d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5683345
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1325425}