Fix CleanupOrphanedBundlesTest.CleanUpSuccessfulOnSessionStart test
Before this CL, this test was not testing the actually intended
behavior. The original test intended to test the orphaned bundle
cleanup by creating the user profile, simulating the orphaned
bundle by creating a directory in the profiles home directory,
performing the login, and finally checking that the directory does
not exist anymore.

However, the orphaned bundle simulation uses the current global profile
directory, which is not updated at the time of bundle simulation (i.e.
after user creation and before login). Hence the check at the end of
the test that checks whether the directory is not there, will always
pass.

This CL fixes that behavior by listening to the profile manager and
creating the directory right after profile creation. This is needed
so that directory can be created right before the
IsolatedWebAppPolicyManager is initialized and the cleanup command
is run.

Note that the production code and the other tests are working as
intended.

Change-Id: I0248c43f6935514b55698b6c9420210dc425fef1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5683484
Reviewed-by: Robbie McElrath <rmcelrath@chromium.org>
Commit-Queue: Simon Hangl <simonha@google.com>
Cr-Commit-Position: refs/heads/main@{#1325429}