Spanify and bounds-check base::circular_deque
Add a range-based constructor, tagged with `base::from_range_t` as a
backport of the C++23 `std::from_range_t`.

Add a range-based assign, named assign_range(), matching std vector and
string in C++23.

Then mark the iterator-pair-based constructor and assign() as
UNSAFE_BUFFER_USAGE.

Change VectorBuffer (the internal class) to stop allowing dereference of
the pointer one past the allocation as this is UB, and CHECK for this
instead. Then stop using derefences in circular_deque that are used to
construct pointers (as we were doing this with the end pointer too and
thus CHECKing). Use pointer arithmetic instead which is valid to
construct the end pointer without invoking UB.

Write SAFETY comments and UNSAFE_BUFFERS() on the pointer arithmetic
inside circular_deque. Move to using spans of the VectorBuffer instead
in most cases.

Make the iterators check for moving out of bounds and dereferencing
end(). Stop going through a pointer to the deque to do iterator
operations to make things easier on the optimizer. However this means
we must reconstruct iterators internally when we change the begin_ or
end_ of the circular_deque, so some minor reordering was needed
internally. This doesn't affect external users as their iterators
would be considered invalidated when begin_ or end_ move anyways.

The binary size regressions are entirely due to the PGO profile being
incorrect after this change: https://crbug.com/344608183#comment7

Bug: 40284755
Fuchsia-Binary-Size: Regression is a temporary PGO artifact
Binary-Size: Regression is a temporary PGO artifact
Change-Id: I20e74d02901387d75aa611ea1e7e98a699a6265a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5655271
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1325442}