[SHv2] Add a missing check to the browser test
The browser test RevokeAllContentSettingTypes checks if after all
permissions were allowed and the auto-revocation threshold has been
passed (i.e. 70 days have past), all of them are correctly auto-revoked.

For that we need to check not only that there are revoked permissions
list for one origin and the lengths of revoked permissions list is the
same as of allowed permissions list, but also that the content of these
two lists are the same too.

This CL adds an assertion that contents of allowed_permission_types and
revoked_permission_types lists are the same.

Bug: 40267370
Fixed: b/350177420
Change-Id: I975cda3d3d07fa28434c0236a2bae56b064a4f72
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5666682
Reviewed-by: Side YILMAZ <sideyilmaz@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Tom Van Goethem <tov@chromium.org>
Commit-Queue: Aisulu Rakhmetullina <aisulu@google.com>
Cr-Commit-Position: refs/heads/main@{#1325470}