[Uno-D] Fix SigninPending avatar icon in RTL
Fix is to get the mirrored value of the bounds to paint the dotted
circle at the right spot.
Adaptation is also needed due to enlarging the image for other purposes.

No tests are currently added for this, there is another on going CL
that targets on adding Pixel tests, a test for RTL should be added as
well there.

Fixed: b:350962399
Change-Id: I3b6b99772d11b3193796dc316bc6c3744f207eba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5691652
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Ryan Sultanem <rsult@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1325576}