[IOSSignout] Enabling finch config and fixing egtest accordingly
This CL updates `testMigrateSyncToSignin_ManagedAccount` and
`testSignOutWithManagedAccountFromNoneSyncingAccount` to run for the disabled feature only. The
feature needs to be disabled because a differnt dialog is shown once it is enabled.

A new test is also added (similar to `testMigrateSyncToSignin_ManagedAccount`) to verify the
expected behaviour with the feature enabled.

The `saveExamplePasswordTo*` methods also save the form's `date_created` as the timestamp is
needed for partial data clearing on signout.

Bug: 351184626
Change-Id: I136c9173ac314d9b382195ceb3c1d845008095c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5678145
Commit-Queue: Salma Elmahallawy <esalma@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1325606}