[Traffic Annotations] Update CEC annotation
An annotation in the Chrome Enterprise Companion is missing some fields.
This was missed during presubmit, because the CEC is not part of the
`traffic_annotation_auditor_dependencies` target.

Add the CEC `traffic_annotation_auditor_dependencies`, and add the
missing fields to the annotation.

Bug: 332808600
Change-Id: Iaee952d9080800de999dfdefb61c391b1a782204
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5685792
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Reviewed-by: Noah Rose Ledesma <noahrose@google.com>
Commit-Queue: Nicolas Ouellet-Payeur <nicolaso@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1325619}