tray: Fix yellow A11y ring incorrect position
When the ChromeVox is enabled, we tab through any detailed view from the
Quick Settings, we found that after we scroll down/up to make the focus
ring and the yellow a11y ring outside of the visible rect, when we press
tab again, the scroll view will automatically scrolled to move the blue
focus ring to the visible rect; however, the yellow focus ring doesn't
followed this change. Thus, there is an offset between the blue focus
ring and the yellow a11y ring.

This CL will try to resolve this bug.

Bug: b/343795715
Test: Manual
Change-Id: I820f700570b39af06fb99a052312e75e58d02baf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5692274
Commit-Queue: Hongyu Long <hongyulong@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Reviewed-by: Richard Chui <richui@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1325620}