Fix potential footgun in UpdateAndGetSessionEnded
UpdateAndGetSessionEnded triggers a call to ProcessEvents which *could*
theoretically cause an AV by referencing a null or uninitialized
input_helper_. This doesn't generally seem to happen because we are
usually destroying the OpenXr wrapper at the same time that we call
Uninitialize, but this can lead to a safer overall pattern.

Also adds a few cases where we early exit if we are no longer
initialized.

Change-Id: I3fe9ee861285a56c77d9aad0efe3b788408aad79
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5685337
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Brandon Jones <bajones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1325659}