Read Aloud: fix how some observers are removed
The "this::..." syntax for passing a method as a callback creates a new
object each time it is used, so calling addObserver(this::aaa) and then
removeObserver(this::aaa) won't actually remove the original. This CL
fixes a few cases where we made this mistake.

Also a few ReadAloudController members are made final, unrelated to
the bug.

Bug: 352204614
Change-Id: I4b2b274b8d90b2ef3efd09381181111b45d84b92
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5692519
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Ian Wells <iwells@chromium.org>
Reviewed-by: Basia Zimirska <basiaz@google.com>
Cr-Commit-Position: refs/heads/main@{#1325676}