Fix crash in OOP crash handler
The gist is that the crash handler uses a static structure for
providing file name and version info. If the info is read while
the handler is crashing then that can cause another crash.

Given that the plan is to move to Crashpad, the simplest workaround
is to remove the static structures and use compile-time values.

Bug: 350725178
Change-Id: I9a768787107a9ece3bda982c799157a2318e079f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5690761
Reviewed-by: Lambros Lambrou <lambroslambrou@chromium.org>
Commit-Queue: Joe Downing <joedow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1325685}